% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howmanyballs.R
\name{howmanyballs}
\alias{howmanyballs}
\title{How many balls?}
\usage{
howmanyballs(x, y, z)
}
\arguments{
\item{x}{length of room in meters}

\item{y}{width of room in meters}

\item{z}{hieght of room in meters}
}
\description{
This function gives a conservative estimate of how many ball-pit balls will fit in your room.
Room dimensions are in meters.
Function assumes simple cubic arrangement of balls of 7.6cm diameter.
You could probably fit more balls, if you want.
}
\examples{
howmanyballs(10, 12, 8)
}
\keyword{ballpit}
\keyword{balls}
\keyword{room}
